Improved the image reader, it now takes 20 milliseconds to load an icons, spend in loading the data from sd card.
Changed the softbuttons code to use callbacks for the button function, this makes it easier to change the button function with the need for some nested switch/if statements.

Tried creating a PWM driven backlight but can's get pwm working at the moment.

Switch backlight off when loading a new page, this hides the slow updating.